+++ example/test_driver/app_test.dart
@@ -15,7 +15,7 @@ void main()
 
     tearDownAll(() async {
       if (driver != null) {
-        driver.close();
+        await driver.close();
       }
     });